Control.Monad.Par.RNG

Description

This module defines another Par-related class to capture the random number generation capability.

The rand operation provides deterministic parallel random number generation from within a Par monad.

Most likely one will simply use the ParRand the instance provided in this file, which is based on a state transformer carrying the random generator.

Documentation

class ParRand p where

A ParRand monad is a Par monad with support for random number generation..

Minimal complete definition

rand

Methods

rand :: Random a => p a

randInt :: p Int

runParRand :: ParFuture fut p => (p a -> a) -> StateT StdGen p a -> IO a

type ParRandStd par a = StateT StdGen par a

A convenience type for the most standard